chore: breaking out models
first attempt at breaking out models into a schema, ideally these would
be their own files as it moves forwards, but i put them all in one place
for the moment to minimize the changes required for review. each break
out would likely be a simple chore style pr in the future.
