+++ api/schemas.py
@@ -1,0 +1,122 @@
+from typing import List, Optional, Dict, Any, Literal
+from pydantic import BaseModel, Field
+
+# Wiki related models
+class WikiPage(BaseModel):
+    """
+    Model for a wiki page.
+    """
+    id: str
+    title: str
+    content: str
+    filePaths: List[str]
+    importance: str  # Should ideally be Literal['high', 'medium', 'low']
+    relatedPages: List[str]
+
+class ProcessedProjectEntry(BaseModel):
+    id: str  # Filename
+    owner: str
+    repo: str
+    name: str  # owner/repo
+    repo_type: str  # Renamed from type to repo_type for clarity with existing models
+    submittedAt: int  # Timestamp
+    language: str  # Extracted from filename
+
+class RepoInfo(BaseModel):
+    owner: str
+    repo: str
+    type: str
+    token: Optional[str] = None
+    localPath: Optional[str] = None
+    repoUrl: Optional[str] = None
+
+class WikiStructureModel(BaseModel):
+    """
+    Model for the overall wiki structure.
+    """
+    id: str
+    title: str
+    description: str
+    pages: List[WikiPage]
+
+class WikiCacheData(BaseModel):
+    """
+    Model for the data to be stored in the wiki cache.
+    """
+    wiki_structure: WikiStructureModel
+    generated_pages: Dict[str, WikiPage]
+    repo_url: Optional[str] = None  # compatible for old cache
+    repo: Optional[RepoInfo] = None
+    provider: Optional[str] = None
+    model: Optional[str] = None
+
+class WikiCacheRequest(BaseModel):
+    """
+    Model for the request body when saving wiki cache.
+    """
+    repo: RepoInfo
+    language: str
+    wiki_structure: WikiStructureModel
+    generated_pages: Dict[str, WikiPage]
+    provider: str
+    model: str
+
+class WikiExportRequest(BaseModel):
+    """
+    Model for requesting a wiki export.
+    """
+    repo_url: str = Field(..., description="URL of the repository")
+    pages: List[WikiPage] = Field(..., description="List of wiki pages to export")
+    format: Literal["markdown", "json"] = Field(..., description="Export format (markdown or json)")
+
+# Model configuration related models
+class Model(BaseModel):
+    """
+    Model for LLM model configuration
+    """
+    id: str = Field(..., description="Model identifier")
+    name: str = Field(..., description="Display name for the model")
+
+class Provider(BaseModel):
+    """
+    Model for LLM provider configuration
+    """
+    id: str = Field(..., description="Provider identifier")
+    name: str = Field(..., description="Display name for the provider")
+    models: List[Model] = Field(..., description="List of available models for this provider")
+    supportsCustomModel: Optional[bool] = Field(False, description="Whether this provider supports custom models")
+
+class ModelConfig(BaseModel):
+    """
+    Model for the entire model configuration
+    """
+    providers: List[Provider] = Field(..., description="List of available model providers")
+    defaultProvider: str = Field(..., description="ID of the default provider")
+
+class AuthorizationConfig(BaseModel):
+    code: str = Field(..., description="Authorization code")
+
+# Chat API models
+class ChatMessage(BaseModel):
+    role: str  # 'user' or 'assistant'
+    content: str
+
+class ChatCompletionRequest(BaseModel):
+    """
+    Model for requesting a chat completion.
+    """
+    repo_url: str = Field(..., description="URL of the repository to query")
+    messages: List[ChatMessage] = Field(..., description="List of chat messages")
+    filePath: Optional[str] = Field(None, description="Optional path to a file in the repository to include in the prompt")
+    token: Optional[str] = Field(None, description="Personal access token for private repositories")
+    type: Optional[str] = Field("github", description="Type of repository (e.g., 'github', 'gitlab', 'bitbucket')")
+
+    # model parameters
+    provider: str = Field("google", description="Model provider (google, openai, openrouter, ollama, bedrock, azure)")
+    model: Optional[str] = Field(None, description="Model name for the specified provider")
+
+    language: Optional[str] = Field("en", description="Language for content generation (e.g., 'en', 'ja', 'zh', 'es', 'kr', 'vi')")
+    excluded_dirs: Optional[str] = Field(None, description="Comma-separated list of directories to exclude from processing")
+    excluded_files: Optional[str] = Field(None, description="Comma-separated list of file patterns to exclude from processing")
+    included_dirs: Optional[str] = Field(None, description="Comma-separated list of directories to include exclusively")
+    included_files: Optional[str] = Field(None, description="Comma-separated list of file patterns to include exclusively")
